🚧 Refactor ticket creation with new ticket worker
This change reorganizes the ticket management functionality into an API-oriented structure, adding capabilities to handle ticket retrieval and processing through API Gateway and support for SQS-based event handling.
