+++ functions/functions/tickets/api/sls.tickets.function.yml
@@ -1,28 +1,21 @@
 Tickets:
   runtime: python3.11
-  handler: functions/tickets/lambda_tickets.lambda_handler
+  handler: functions/tickets/api/lambda_tickets.lambda_handler
   name: "${sls:stage}-${self:service}-tickets"
   package:
     patterns:
       - '!**/**'
-      - "functions/tickets/**"
+      - "functions/tickets/api/**"
       - "_pydantic/**" # this requires the pydantic layer also
       - "_shared/**"
   environment:
       STAGE_NAME: ${sls:stage}
       ORG_TABLE_NAME_TEMPLATE: ${self:custom.dynamodb_table_format}
-      TICKET_QR_JWT_SECRET: ${ssm:/danceengine/${sls:stage}/jwt_secret}
   layers:
       - !Ref UtilsLambdaLayer
       - !Ref PydanticLambdaLayer # uncomment if you need pydantic layer
       - !Ref AuthLambdaLayer
   events:
-    - eventBridge:
-        pattern:
-          source:
-            - ${param:default_event_bus_source}
-          detail-type:
-            - "checkout.completed"
     - httpApi:
         path: /{organisation}/{event}/tickets
         method: get
