+++ functions/functions/tickets/create/lambda_create_tickets.py
@@ -9,8 +9,6 @@ from datetime import datetime, timezone
 
 ## installed packages
 from pydantic import ValidationError # layer: pydantic
-from boto3.dynamodb.conditions import Key
-from botocore.exceptions import ClientError
 from ksuid import KsuidMs # layer: utils
 import jwt
 
@@ -19,7 +17,6 @@ sys.path.append(os.path.dirname(__file__))
 from _shared.parser import parse_event
 from _shared.DecimalEncoder import DecimalEncoder
 from _shared.helpers import make_response
-from _pydantic.models.tickets_models import TicketListResponse
 from _pydantic.models.models_extended import TicketModel, TicketChildModel, CustomerModel
 from _pydantic.EventBridge import trigger_eventbridge_event, EventType, Action, EventBridgeEvent, EventBridgeEventDetail # pydantic layer
 from _pydantic.dynamodb import transact_upsert # pydantic layer
@@ -231,89 +228,16 @@ def create_ticket(request_data: EventBridgeEvent, organisation_slug: str, actor:
         logger.error(traceback.format_exc())
         return make_response(500, {"message": "Something went wrong."})
     
-def get_single_ticket(organisationSlug: str,  eventId: str,  ticketId: str, public: bool = False, actor: str = "unknown"):
-    TABLE_NAME = ORG_TABLE_NAME_TEMPLATE.replace("org_name", organisationSlug)
-    table = db.Table(TABLE_NAME)
-    logger.info(f"Getting ticket for {eventId} of {organisationSlug} from {TABLE_NAME}")
-    blank_model = TicketModel(ksuid=ticketId, parent_event_ksuid=eventId, name="blank", organisation=organisationSlug, name_on_ticket="blank", customer_email="blank", email="blank", includes=[])
-
-    try:
-        ticket = blank_model.query_gsi(
-            index_name="gsi2",
-            table=table, 
-            key_condition=Key("gsi2PK").eq(f'{blank_model.gsi2PK}'), 
-            assemble_entites=True
-        )
-        logger.info(f"Found ticket for {organisationSlug}: {ticket}")
-    except ClientError as e:
-        if e.response['Error']['Code'] == 'ResourceNotFoundException':
-            logger.error(f"Ticket not found for {organisationSlug}: {e}")
-            return None
-        else:
-            raise
-    except ValueError as e:
-        logger.error(f"Ticket not found for {organisationSlug}: {e}")
-        return None
-    
-    except Exception as e:
-        logger.error(f"DynamoDB query failed to get ticket for {organisationSlug}: {e}")
-        raise Exception
-    
-    return ticket if ticket else None
-
-def get_tickets(organisationSlug: str,  eventId: str, public: bool = False, actor: str = "unknown"):
-    TABLE_NAME = ORG_TABLE_NAME_TEMPLATE.replace("org_name",organisationSlug)
-    table = db.Table(TABLE_NAME)
-    logger.info(f"Getting Tickets for {eventId} of {organisationSlug} from {TABLE_NAME}")
-    blank_model = TicketModel(ksuid="blank", parent_event_ksuid=eventId, name="blank", organisation=organisationSlug, name_on_ticket="blank", customer_email="blank", email="blank", includes=[])
-
-    try:
-        tickets = blank_model.query_gsi(
-            table=table,
-            index_name="gsi1",
-            key_condition=Key("gsi1PK").eq(blank_model.gsi1PK) & Key("gsi1SK").begins_with(f"{blank_model.gsi1SK.split('#')[0]}#"),
-        )
-        logger.info(f"Found tickets for {eventId} of {organisationSlug}: {tickets}")
-    except Exception as e:
-        logger.error(f"DynamoDB query failed to get tickets for {organisationSlug}: {e}")
-        raise Exception
-    
-    #! temporary fix this needs review
-    if isinstance(tickets, TicketModel):
-        tickets = [tickets]    
-
-    return tickets
-
 def lambda_handler(event, context):
     logger.info("Received event: %s", json.dumps(event, indent=2, cls=DecimalEncoder))
 
-    http_method = event.get('requestContext', {}).get("http", {}).get("method")
-    
-    if http_method:
-        logger.info("Triggered by API Gateway")
-        parsed_event = parse_event(event)
+    if event.get("Records") and event.get("Records", [])[0].get("eventSource") == "aws:sqs":
+        logger.info("Triggered by SQS")
 
-        organisationSlug = event.get("pathParameters", {}).get("organisation")
-        ticketId         = event.get("pathParameters", {}).get("ksuid")
-        eventId          = event.get("pathParameters", {}).get("event")
-        is_public        = event.get("rawPath", "").startswith("/public")
-        actor            = event.get("requestContext", {}).get("accountId", "unknown")
+        parsed_message = parse_event(event.get("Records", [])[0].get("body"))
 
-        if http_method == "GET":
-            logger.info(f"{organisationSlug}:{eventId}:{ticketId} - Getting ticket(s)")
-            response_cls = TicketListResponse
-
-            tickets = [get_single_ticket(organisationSlug, eventId, ticketId, is_public, actor)] if ticketId else get_tickets(organisationSlug, eventId, is_public, actor)
-            if tickets is None:
-                return make_response(404, {"message": "Ticket(s) not found."})
-            
-            resposne = response_cls(tickets=tickets)
-            return make_response(200, resposne.model_dump(mode="json", exclude_none=True))        
-    else:
-        logger.info("Triggered by EventBridge")
-
-        type = event.get("detail-type", {})
-        organisationSlug = event.get("detail", {}).get("organisation", {})
+        type = parsed_message.get("detail", {}).get("type")
+        organisationSlug = parsed_message.get("detail", {}).get("organisation")
 
         if type == "checkout.completed":
             logger.info("Handling checkout.completed event.")
@@ -322,4 +246,4 @@ def lambda_handler(event, context):
             return create_ticket(validated_request, organisationSlug)
         else:
             logger.warning(f"Received event type I can not currently process: {type}")
-            return make_response(400, {"message": f"Unhandled event type: {type}"})
+            return make_response(400, {"message": f"Unhandled event type: {type}"})
