+++ functions/serverless.yml
@@ -179,7 +179,8 @@ layers:
 
 functions:
   Email: ${file(functions/email/sls.email.function.yml):Email}
-  Tickets: ${file(functions/tickets/sls.tickets.function.yml):Tickets}
+  CreateTickets: ${file(functions/tickets/create/sls.create_tickets.function.yml):CreateTickets}
+  Tickets: ${file(functions/tickets/api/sls.tickets.function.yml):Tickets}
   EventbridgeCheckout: ${file(functions/checkout/sls.checkout.function.yml):EventbridgeCheckout}
   Checkout: ${file(functions/checkout/sls.checkout.function.yml):Checkout}
   Bundles: ${file(functions/bundles/sls.bundles.function.yml):Bundles}
